requirements.txt
librosa==0.9.2
numba==0.56.0
numpy==1.19.5
nltk==3.7
pandas==1.2.5
pytokenizations==0.7.2
scikit-learn==0.19.2
scipy==1.7.0
seaborn==0.11.1
torch==1.9.0
torch-struct==0.5
torchaudio==0.9.0
torchvision==0.10.0
transformers==4.3.0
transliterate==1.10.2
